Demonstrating level of MVC understanding and PHP mastery
Was supplied code that combined Database.php, Index.php,
Database_error.php, and mgsFunctions into one file and instead of
functions was supplied the code within mgsFunctions in a raw format. I
took the code and separated it into an MVC model and created functions
to perform the tasks of the raw data and isolated the code.